requirements.txt
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
et_xmlfile==2.0.0
idna==3.10
msal==1.33.0
openpyxl==3.1.5
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
urllib3==2.5.0
